Reads /proc/*/maps after successful exec
Before execve() "returns", the kernel will have mapped ld-linux.so to
dynamically link the ELF executable. That executable may also be
different from execve()'s first argument, if that argument is a script
file with a shebang (#!/some/interpreter).

However, both the interpreter and ld-linux.so would appear in the maps
pseudo-file.