Handles symbolic links at the 'trace' step
Previously, only the symlink itself would be considered, assigned to a
package and written to the configuration. The 'pack' step would then see
the links that the path traverses, and the final target, and pack them
with the link.

This was wrong, because a symlink's target can come from a package
itself, and the fact that these files were to be handled separately
without appearing in the configuration posed other problems (when
unpacking).

The tracer now adds the symlinks and targets to the list of files to the
list of files that is read from the database. They still don't appear
there, but are in the configuration from the start. They will be
categorized in the packages, and other steps do not need to worry about
them (if a symlink appears in the configuration, we can just assume that
the target is also listed).