Fixes bug if new processes resumes first
When we fork/clone, the new process might start its execution first (and
tracer sees it get a SIGTRAP). In this case, we wait for the fork call
to return before we insert things into the database and resume it
(previously, a unknown process would be created for it, plus a duplicate
when fork actually returned. That was bad).