Adds additional_patterns section
With this, we have two types of configuration files: the ones that the
user is supposed to edit, that the packer reads; and the one that gets
stored in a .rpz pack, that the unpackers read.

The only difference is that the packed one has been processed by the
software last and is "canonical", i.e. reflect what is in the pack; the
other might list missing files, and have entries in additional_patterns
that have not been expanded yet.

TODO: remove missing files from canonical config file.