Changes plugin entry_point protocol
This changes the API with which plugins register with reprounzip.

The name of the entry_point is now used as the name of the reprounzip
subcommand (e.g. 'reprounzip graph') and the declared entry_point
function takes the subparser directly as parameter. The function's
docstring is used as help message.

These setup functions might return either None or a dictonary containing
information on the command (this will be useful later, e.g. to display
which plugins are compatible).