Fixes the packer failing on temporary directories
In addition to the configuration file, the packer also reads the trace
to pack directories in which files were created or that were used as
working directory of processes. However these might have been temporary
directories that no longer exist; in this case, an OSError really
shouldn't be allowed to propagate.