Just bail out if using Python < 2.7.3
It has been confirmed that reprozip will NOT work, as it is unable to
read trace.sqlite3. 2.7.3 was released in Aug 2012 and is included in
Debian 7.0 Wheezy, so people shouldn't be using these versions anyway.

Working around that bug would mean changing the way I store arrays,
which would need some kind of escaping of a chosen character, and would
break packfile compatibility.