Runs selected command via 'vagrant run'
Running commands via 'vagrant run' no longer uses the 'script.sh' file.

'run' now take the run number as parameter, so you can select which
command to run. Defaults to 0 if there is only one, and selects every
run if not specified and --cmdline is not used.

The --cmdline option allows to view the command-line of the selected run
(if passed without arguments) or to specify a new one. The uid/gid and
working directory of the original run are kept.