Lists parent directories in config
In addition to the files, the packer also includes the directories which
are used as CWD or in which files are created. However, these were
identified when packing, not when tracing, and thus didn't appear in the
config file.

This patch makes it happen when tracing and puts them in the config
file, which the packer then reads.