Use PTRACE_EVENT_EXEC instead of syscall return
Previously, I was relying on the syscall-return-stop out of execve() to
detect that an exec had happened. This had a number of drawbacks:
* Can't be sure which thread executed (although race conditions here are
  not common)
* Need to read orig_eax on syscall-return-stop to know which syscall we
  are returning from, although in normal circumstances we should know
  (it's the one we entered). Also, that value might apparently be set
  wrong...
* Needed to work around i386/x86_64 transitions, as Linux could report a
  thread going out of "syscall 59", number for execve in x86_64,
  although the new program is running in i386 mode

We now react to PTRACE_EVENT_EXEC and do the work there. We then set the
flag PROCFLAG_EXECD, so that we know the next syscall-in/out-stop is in
fact a return from execve, no matter what Linux reports (and whether or
not the thread had entered a syscall before).