Use PTRACE_EVENT_FORK, ... instead of syscall out
Previously, I was only using the syscall-return-stop out of fork(),
vfork() or clone() to detect that a successful process creation had
happened. This was a problem because syscall return can happen either
before or after the process is created.

* When we return first, we created a new PROCSTAT_ALLOCATED that will be
  seen later (then options get set and it turns into PROCSTAT_ATTACHED).
* When it appears first, an unknown process was created in
  PROCSTAT_UNKNOWN and stays STOPped until the syscall returns, set it
  to it to PROCSTAT_ATTACHED and resumed it.

This was already a workaround, but worked. Problem is that vfork()
doesn't actually return before the new process appeared and run until it
dies or calls execve(), which couldn't happen because the syscall
resumes it; this meant deadlock on any vfork() call.

We now react to PTRACE_EVENT_{FORK,VFORK,CLONE} instead, which always
happens before the new process appears. We can get rid of the two
different handling depending on the random order chosen by the kernel,
and vfork won't deadlock anymore.