Workaround for X11 and remote Docker daemons
This detects if the Docker daemon is a remote machine and displays a
warning.

It also adds --tunneled-x11 which makes the Docker container connect to
its local machine (the server) instead of trying to contact the local IP
address (the machine reprounzip is running on). In this case, the user
is responsible for creating the SSH tunnel (which will probably require
'GatewayPorts yes' in the server's SSHd configuration).