Introduce "parameters"
These are application settings that can change between ReproZip
releases, for example as software repositories move or new operating
systems are released.

The parameters are stored in a simple JSON file, that will be downloaded
from the web. A version of that file will be bundled with each release
(though without internet connection, some unpackers won't work, as they
need to download busybox/rpzsudo).