X11Handler#env_fixes(), parse_environment_args()
Those methods return the environment variables to be set and unset, for
X11 and the command-line arguments respectively, instead of directly
returning the changed environment like the current methods do
(X11Handler#fix_env() and fixup_environment()).

This is useful for Docker, which merges the changes with the original
environment internally with `env`.