Correctly set PROCFLAG_OPEN_EXIST from openat()
When handling open(), the file is checked for existance before the
syscall, and that information is used when coming out of the syscall to
record read-write as simply write if the file didn't exist.

However openat() was only handled after the syscall, and ran the handler
for open() that expected this flag to have been set.

This adds the call to the before-open() handler for openat() too to set
that flag.